fix review comments 03/08/20
